add Calculate

+++ MLP/src/my/mlp/IonosphereParser.java
@@ -38,24 +38,24 @@ public class IonosphereParser {
 		fileReader.close();
 	}
 	
-	public List<HashMap<String, float[]>> getTraindata(){
+	public List<HashMap<String, double[]>> getTraindata(){
 		
 		output = new ArrayList<String>();
 		
 		int num = (int) (iono.size() * 0.75);
 		
-		List<HashMap<String,float[]>> list = new ArrayList<HashMap<String, float[]>>();
+		List<HashMap<String,double[]>> list = new ArrayList<HashMap<String, double[]>>();
 		
 		for(int i = 0; i < num; i++)
 		{
-			HashMap<String, float[]> hm = new HashMap<String, float[]>();
+			HashMap<String, double[]> hm = new HashMap<String, double[]>();
 			String[] line = iono.get(i).split(",");
 			if(!output.contains(line[0]))
 				output.add(line[0]);
-			float[] par = new float[line.length-1];
+			double[] par = new double[line.length-1];
 			for(int j = 0; j < line.length-1; j++)
 			{
-				par[j] = Float.valueOf(line[j]);
+				par[j] = Double.valueOf(line[j]);
 			}
 			hm.put(line[line.length -1], par);
 			list.add(hm);
@@ -64,21 +64,21 @@ public class IonosphereParser {
 		return list;
 	}
 	
-public List<HashMap<String, float[]>> getTestdata(){
+public List<HashMap<String, double[]>> getTestdata(){
 		
 		int num = (int) (iono.size() * 0.75);
 		int start = iono.size() - num;
 		
-		List<HashMap<String,float[]>> list = new ArrayList<HashMap<String, float[]>>();
+		List<HashMap<String,double[]>> list = new ArrayList<HashMap<String, double[]>>();
 		
 		for(int i = start; i < iono.size(); i++)
 		{
-			HashMap<String, float[]> hm = new HashMap<String, float[]>();
+			HashMap<String, double[]> hm = new HashMap<String, double[]>();
 			String[] line = iono.get(i).split(",");
-			float[] par = new float[line.length-1];
+			double[] par = new double[line.length-1];
 			for(int j = 0; j < line.length - 1; j++)
 			{
-				par[j] = Float.valueOf(line[j]);
+				par[j] = Double.valueOf(line[j]);
 			}
 			hm.put(line[line.length - 1], par);
 			list.add(hm);
